Simplify test coverage architecture
- Remove unnecessary useTestCoverage hook (overkill for single component)
- Move fetch logic directly into TestCoverageCard component
- Simplify component props - no longer needs external data/loading/error props
- Maintain same functionality with less abstraction
- Keep TestTube import for header icons in different states
